demo/wechatpusher/requirements.txt
python-dotenv
requests
pandas
schedule
langchain
langchain-openai
langchain-community
numpy==1.26